Improvements to workers and job list

+++ rooibos/workers/views.py
@@ -15,12 +15,17 @@ def joblist(request):
         jobs = JobInfo.objects.filter(owner=request.user)
     
     if request.method == "POST":
-        ids = request.POST.getlist('r')
-        jobs.filter(id__in=ids, completed=True).delete()
+        if request.POST.get('remove'):
+            ids = request.POST.getlist('r')
+            jobs.filter(id__in=ids, completed=True).delete()
+        else:
+            for k, v in request.POST.iteritems():
+                if k.startswith('run-'):
+                    JobInfo.objects.get(id=k[4:]).run()
+        
         return HttpResponseRedirect(reverse('workers-jobs'))
     
     return render_to_response("workers_jobs.html",
                               {'jobs': jobs,
-                               'owner': request.user if not request.user.is_superuser else None,
                                },
                               context_instance=RequestContext(request))
